pyproject.toml
[project]
name = "foggypark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=0.3.26",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "watchdog>=6.0.0",
    "weaviate-client>=4.16.0",
]
